+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Support OpenTool Spec 2.0, add stream Call
+- separate server and tool
+
 ## 1.0.0
 
 - Initial version.